extract response mod, move QueryResponse there
