+++ scylla/src/execution/iterator.rs
@@ -24,7 +24,7 @@ use crate::execution::ExecutionProfileInner;
 
 use crate::cluster::ClusterData;
 use crate::cluster::{Node, NodeRef};
-use crate::connection::{Connection, NonErrorQueryResponse, QueryResponse};
+use crate::connection::Connection;
 use crate::execution::history::{self, HistoryListener};
 use crate::execution::load_balancing::{self, RoutingInfo};
 use crate::execution::retries::{QueryInfo, RetryDecision, RetrySession};
@@ -36,6 +36,7 @@ use crate::frame::{
     },
     value::SerializedValues,
 };
+use crate::response::{NonErrorQueryResponse, QueryResponse};
 use crate::statement::Consistency;
 use crate::statement::{prepared_statement::PreparedStatement, query::Query};
 use tracing::{trace, trace_span, warn, Instrument};
