+++ scylla/src/response/mod.rs
@@ -1,0 +1,2 @@
+mod query_response;
+pub(crate) use query_response::{NonErrorQueryResponse, QueryResponse};
